添加版本检测，优化逻辑

+++ bin/nosync.js
@@ -7,12 +7,14 @@ const program = require('commander');
 const chalk = require('chalk');
 const inquirer = require('inquirer');
 const ora = require('ora');
+const checkForUpdate = require('update-check');
 
 const pwd = process.cwd(); // 当前目录
+const pkg = require('../package');
 const spinner = ora('转化中 🐢 ...\n');
 
 program
-  .version(require('../package').version, '-v, --version')
+  .version(pkg.version, '-v, --version')
   .option('-f, --folder [name]', '禁止同步的[文件夹名]', 'node_modules')
   .option('-g, --git [boolean]', '是否自动添加 .gitignore')
   .parse(process.argv);
@@ -20,8 +22,35 @@ program
 const basePath = path.join(pwd, program.folder); // 用户期望不同步的文件夹的路径，默认是当前目录下的 node_modules
 const nosyncPath = basePath + '.nosync'; // 同名的 nosync 型文件夹的路径
 
+// 检测 npm 版本，提示用户更新
+checkForUpdate(pkg).then(update => {
+  console.log(chalk.yellow(`\n检测到最新的版本：${pkg.version}`), chalk.white('-->'), chalk.green(update.latest), chalk.yellow('，建议您稍后更新！\n'));
+}).catch(() => {
+  console.log(chalk.red('\n检测更新失败，请查看您的网络状况！\n'));
+}).finally(() => {
+  // 无论版本检测成功与否，继续执行业务逻辑
+  checkPwd()
+    .then(() => createNosyncFolder())
+    .then(res => {
+      // 为 nosync 文件夹制作替身
+      fs.symlinkSync(nosyncPath, basePath, 'file');
+      // 结束进程指示器
+      spinner.stop();
+      // 如果 指定的 nosync 文件夹是 node_modules 并且 之前不存在 node_modules，则提示用户安装
+      if (program.folder === 'node_modules' && res.install) {
+        install();
+      } else {
+        // 输出成功提示并提示是否将文件夹添加到 .gitignore
+        addGitignore();
+      }
+    })
+    .catch(() => {
+      process.exit();
+    });
+});
+
 /**
- * 检测项目当前位置是否在 iCloud 中（用户并不一定会将桌面和文稿也同步到 iCloud，给个友情提示）
+ * 检测项目当前位置是否在 iCloud 目录中（有些用户可能使用 iCloud 同步桌面和文稿，给个友情提示）
  */
 function checkPwd() {
   return new Promise((resolve, reject) => {
@@ -32,10 +61,10 @@ function checkPwd() {
         .prompt([
           {
             type: 'confirm',
-            message: '检测到你当前的项目不在 iCloud 中，是否继续？',
+            message: '您当前的项目不在 iCloud 目录中，是否继续？',
             name: 'continue'
-					}
-				])
+          }
+        ])
         .then(answers => {
           answers.continue ? resolve(answers) : reject(answers);
         })
@@ -51,29 +80,53 @@ function checkPwd() {
  */
 function createNosyncFolder() {
   spinner.start();
-  const baseExist = fs.existsSync(basePath); // 期望的路径是否已占用
-  const nosyncExist = fs.existsSync(nosyncPath); // 期望的 nosync 路径是否已占用
-  const baseIsDirectory = baseExist && fs.lstatSync(basePath).isDirectory(); // 期望的路径是否已存在文件夹
+  const baseExist = fs.existsSync(basePath); // 期望的路径是否存在（如果该路径是 SymbolicLink, 则该方法实际检测的是它链接的地址是否已存在）
+  const nosyncExist = fs.existsSync(nosyncPath); // 期望的 nosync 路径是否存在
+  const baseIsDirectory = baseExist && fs.lstatSync(basePath).isDirectory(); // 期望的路径是否是文件夹
 
   return new Promise((resolve, reject) => {
-    if (baseExist && nosyncExist) { // 1. 同时存在，提示用户无需额外操作
+    if (baseExist && nosyncExist) { // 1. 同时存在，如果是 node_modules 则提示用户是否再次安装包，否则退出
       spinner.stop();
-      console.log(chalk.black.bgRed('\n你多虑了!\n'));
-      reject();
+      if (program.folder === 'node_modules') {
+        install();
+      } else {
+        console.log(chalk.yellow(`\n${program.folder} 已不再同步到 iCloud 了，您无需重复执行！\n`));
+        reject();
+      }
     } else if (!baseExist && nosyncExist) { // 2. 只存在 nosync，直接 resolve 后制作替身
+      // basePath 可能存在无效的 SymbolicLink，这种情况的概率很低，但不删除的话会导致之后调用 fs.symlinkSync() 报错
+      try {
+        fs.unlinkSync(basePath);
+        console.log(`已移除无效的 ${program.folder} 快捷方式`);
+      } catch (error) {
+        // do nothing and keep silence
+      }
       resolve({ install: false });
     } else if (baseExist && baseIsDirectory) { // 3. 指定的文件夹已存在，重命名为 nosync 型
       fs.rename(basePath, nosyncPath, (err) => {
         if (err) throw err;
         resolve({ install: false });
       });
-    } else if (baseExist && !baseIsDirectory) { // 4. 之前创建过快捷方式，删除快捷方式并创建空的 nosync 型文件夹
-      fs.unlinkSync(basePath);
-      fs.mkdir(nosyncPath, (err) => {
-        if (err) throw err;
-        resolve({ install: true });
-      });
-    } else { // 5. 都不存在，创建 nosync 型文件夹
+    } else if (baseExist && !baseIsDirectory) { // 4. 指定的是文件而非文件夹
+      if (program.folder === 'node_modules') {
+        fs.unlinkSync(basePath);
+        fs.mkdir(nosyncPath, (err) => {
+          if (err) throw err;
+          resolve({ install: true });
+        });
+      } else {
+        fs.rename(basePath, nosyncPath, (err) => {
+          if (err) throw err;
+          resolve({ install: false });
+        });
+      }
+    } else { // 5. basePath 不存在 或者存在无效的 SymbolicLink
+      try {
+        fs.unlinkSync(basePath);
+        console.log(`已移除无效的 ${program.folder} 快捷方式`);
+      } catch (error) {
+        // do nothing and keep silence
+      }
       fs.mkdir(nosyncPath, (err) => {
         if (err) throw err;
         resolve({ install: true });
@@ -92,14 +145,14 @@ function install() {
         type: 'list',
         message: '请选择安装 node_modules 的方式？',
         choices: [
-					'yarn',
-					'npm',
-					'cnpm',
-					'暂不安装'
-				],
+          'yarn',
+          'npm',
+          'cnpm',
+          '暂不安装'
+        ],
         name: 'install'
-			}
-		])
+      }
+    ])
     .then((res) => {
       let command = '';
       switch (res.install) {
@@ -126,7 +179,7 @@ function install() {
  * 添加忽略规则到 .gitignore
  */
 function addGitignore() {
-  console.log(chalk.green(`\n大功告成，文件夹 ${program.folder} 将不再同步到 iCloud 👏 👏 👏\n`));
+  console.log(chalk.green(`\n大功告成，${program.folder} 将不再同步到 iCloud 👏 👏 👏\n`));
 
   // 如果用户已通过命令行指定是否添加到 git，则不再提示
   if (String(program.git) === 'true') {
@@ -143,8 +196,8 @@ function addGitignore() {
         type: 'confirm',
         message: `是否添加 ${program.folder}* 到 .gitignore？`,
         name: 'add'
-			}
-		])
+      }
+    ])
     .then(answers => {
       answers.add && add();
     });
@@ -165,22 +218,3 @@ function addGitignore() {
     }
   }
 }
-
-checkPwd()
-  .then(() => createNosyncFolder())
-  .then(res => {
-    // 为 nosync 文件夹制作替身
-    fs.symlinkSync(nosyncPath, basePath, 'file');
-    // 结束进程指示器
-    spinner.stop();
-    // 如果 指定的 nosync 文件夹是 node_modules 并且 之前不存在 node_modules，则提示用户安装
-    if (program.folder === 'node_modules' && res.install) {
-      install();
-    } else {
-      // 输出成功提示并提示是否将文件夹添加到 .gitignore
-      addGitignore();
-    }
-  })
-  .catch(() => {
-    process.exit();
-  });
